Rename all snake_cased folders under src/elements/screens to camelCase
