+++ app/packs/src/components/elements/screens/ScreenDetails.js
@@ -18,8 +18,8 @@ import PrivateNoteElement from 'src/components/elements/PrivateNoteElement';
 import QuillEditor from 'src/components/QuillEditor';
 import ResearchPlansFetcher from 'src/fetchers/ResearchPlansFetcher';
 import Screen from 'src/models/Screen';
-import ScreenDetailsContainers from 'src/components/elements/screens/analyses_tab/ScreenDetailsContainers';
-import ScreenResearchPlans from 'src/components/elements/screens/research_plans_tab/ScreenResearchPlans';
+import ScreenDetailsContainers from 'src/components/elements/screens/analysesTab/ScreenDetailsContainers';
+import ScreenResearchPlans from 'src/components/elements/screens/researchPlansTab/ScreenResearchPlans';
 import ScreenWellplates from 'src/components/elements/screens/ScreenWellplates';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import UIStore from 'src/stores/alt/stores/UIStore';
