+++ app/packs/src/components/elements/screens/analysesTab/ScreenDetailsContainers.js
