+++ app/packs/src/components/elements/screens/researchPlansTab/EmbeddedResearchPlanDetails.js
